LayoutTests/fast/backgrounds/background-repeat-x-y-parse-expected.txt

This test checks that background-repeat-x/y are not parsed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(document.getElementById('test1')).getPropertyValue('background-repeat') is "repeat"
PASS window.getComputedStyle(document.getElementById('test2')).getPropertyValue('background-repeat') is "repeat"
PASS window.getComputedStyle(document.getElementById('test3')).getPropertyValue('background-repeat') is "repeat-y"
PASS window.getComputedStyle(document.getElementById('test4')).getPropertyValue('background-repeat') is "repeat-x"
PASS successfullyParsed is true

TEST COMPLETE
